create ipCountTimer component file
